refactor(hooks): Split useEdenMutation into specialized hooks
The monolithic `useEdenMutation` hook has been refactored and split into several specialized components to improve modularity, clarify API usage, and enhance maintainability.

The changes include:
-   **`apps/dockstat/src/hooks/eden/types.ts`**: Introduced to centralize shared types for Eden-related hooks.
-   **`apps/dockstat/src/hooks/eden/useBaseEdenMutation.ts`**: Extracts the core `useMutation` logic, toast handling, and query invalidation into a reusable base hook.
-   **`apps/dockstat/src/hooks/eden/useEdenMutation.ts`**: Provides a simplified interface for direct Eden API routes that do not require a route builder with dynamic parameters.
-   **`apps/dockstat/src/hooks/eden/useEdenRouteMutation.ts`**: Provides an interface for Eden API routes that require parameters to build the route function at the time of mutation.
-   **Removed `apps/dockstat/src/hooks/useEdenMutation.ts`**: The old, overloaded hook has been removed.

All existing consumers of the previous `useEdenMutation` have been updated to use the appropriate new specialized hook (`useEdenMutation` or `useEdenRouteMutation`) and their new import paths.

Additionally, corrected query invalidation keys for `pinNavLink` and `unpinNavLink` mutations in `layout.tsx` from `fetchNavLinks` to `fetchAdditionalSettings` to ensure proper cache updates.
