+++ apps/dockstat/src/hooks/plugins/usePluginActions.ts
@@ -2,7 +2,7 @@ import type { ActionConfig, PageTemplate } from "@dockstat/template-renderer"
 import { useCallback, useRef } from "react"
 import type { ResolvedAction } from "@/components/plugins/id/types"
 import { api } from "@/lib/api"
-import { useEdenMutation } from "../useEdenMutation"
+import { useEdenRouteMutation } from "../eden/useEdenRouteMutation"
 import { getValueByPath } from "./utils"
 
 type ActionResult = {
@@ -39,7 +39,7 @@ export function usePluginActions({
   )
 
   // Correct route builder
-  const actionMutation = useEdenMutation({
+  const actionMutation = useEdenRouteMutation({
     mutationKey: ["executeAction", String(pluginId)],
     routeBuilder: ({ actionId }: { actionId: string }) =>
       api.plugins.frontend({ pluginId }).actions({ actionId }).execute.post,
