+++ apps/dockstat/src/layout.tsx
@@ -6,7 +6,7 @@ import { useContext, useEffect, useState } from "react"
 import { Toaster } from "sonner"
 import { ConfigProviderContext } from "@/contexts/config"
 import { PageHeadingContext } from "./contexts/pageHeadingContext"
-import { useEdenMutation } from "./hooks/useEdenMutation"
+import { useEdenMutation } from "./hooks/eden/useEdenMutation"
 import { useEdenQuery } from "./hooks/useEdenQuery"
 import { useGlobalBusy } from "./hooks/useGlobalBusy"
 import { api } from "./lib/api"
@@ -29,13 +29,13 @@ export default function Layout({ children }: { children: React.ReactNode }) {
   const pinMutation = useEdenMutation({
     mutationKey: ["pinNavLink"],
     route: api.db.config.pinItem.post,
-    invalidateQueries: [["fetchNavLinks"]],
+    invalidateQueries: [["fetchAdditionalSettings"]],
   })
 
   const unPinMutation = useEdenMutation({
     mutationKey: ["unPinNavLink"],
     route: api.db.config.unpinItem.post,
-    invalidateQueries: [["fetchNavLinks"]],
+    invalidateQueries: [["fetchAdditionalSettings"]],
   })
 
   useEffect(() => rssFeedEffect(setRamUsage), [])
