+++ apps/dockstat/src/pages/extensions/index.tsx
@@ -3,7 +3,7 @@ import { Button, Card, Input, Select, Toggle } from "@dockstat/ui"
 import { Plus } from "lucide-react"
 import { useState } from "react"
 import { RepoCard } from "@/components/extensions/RepoCard"
-import { useEdenMutation } from "@/hooks/useEdenMutation"
+import { useEdenMutation } from "@/hooks/eden/useEdenMutation"
 import { useEdenQuery } from "@/hooks/useEdenQuery"
 import { usePageHeading } from "@/hooks/useHeading"
 import { api } from "@/lib/api"
